update regex to avoid incorrect match
with the original regex, it will match some incorrect classes. For example, if the itemP is ".abc", then ".abc2" will also be considered as a match.
This bug will cause the incorrect CSS text in demo. For example, if you click on <a href="#" class="hbtn hb-fill-middle">Slide Middle</a>, it will show CSS text for class ".hb-fill-middle2:hover"